+++ AlbaMobileAndroid/src/main/java/es/albainformatica/albamobileandroid/oldcatalogo/ItemDepartam.kt
@@ -5,7 +5,7 @@ import es.albainformatica.albamobileandroid.ancho_grupo
 import es.albainformatica.albamobileandroid.ponerCeros
 
 
-class ItemDepartam(var grupo: Int, var codigo: Int, var descr: String) {
+class ItemDepartam(var grupo: Short, var codigo: Short, var descr: String) {
     val imagen: String
         get() = "DPT_" + ponerCeros(grupo.toString(), ancho_grupo) +
                 ponerCeros(codigo.toString(), ancho_departamento) + ".jpg"
